Export torchscript (contrary to the branch name)

See merge request aias/slam/nn-slam-benchmark!10

+++ nnsb/vpr_systems/sela/backbone/vision_transformer.py
@@ -209,7 +209,7 @@ class DinoVisionTransformer(nn.Module):
             patch_pos_embed.reshape(
                 1, int(math.sqrt(N)), int(math.sqrt(N)), dim
             ).permute(0, 3, 1, 2),
-            scale_factor=(w0 / math.sqrt(N), h0 / math.sqrt(N)),
+            scale_factor=(float(w0 / math.sqrt(N)), float(h0 / math.sqrt(N))),
             mode="bicubic",
         )
